ai-services/requirements.txt
fastapi==0.103.1
uvicorn[standard]==0.23.2
langchain==0.0.292
langchain-experimental==0.0.30
openai==0.28.1
anthropic==0.3.11
pinecone-client==2.2.4
chromadb==0.4.10
sentence-transformers==2.2.2
transformers==4.33.2
torch==2.0.1
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
nltk==3.8.1
spacy==3.6.1
python-multipart==0.0.6
pydantic==2.3.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
redis==4.6.0
sqlalchemy==2.0.20
alembic==1.11.3
psycopg2-binary==2.9.7
celery==5.3.1
flower==2.0.1
httpx==0.24.1
aiohttp==3.8.5
websockets==11.0.3
pydub==0.25.1
pillow==10.0.0
opencv-python==4.8.0.76
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.15.0
streamlit==1.26.0
pytest==7.4.0
pytest-asyncio==0.21.1
black==23.7.0
flake8==6.0.0
mypy==1.5.1